requirements.txt
platformdirs==4.3.8
pygame==2.6.1
python-dotenv==1.1.1
toml==0.10.2
